# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/ldap/settings": {
            "get": {
                "description": "Retrieves the current LDAP configuration settings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LDAP"
                ],
                "summary": "Get LDAP settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.LDAPSettings"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Updates the LDAP configuration settings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LDAP"
                ],
                "summary": "Update LDAP settings",
                "parameters": [
                    {
                        "description": "LDAP settings",
                        "name": "settings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.LDAPSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/ldap/test": {
            "post": {
                "description": "Tests the LDAP connection with provided credentials",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LDAP"
                ],
                "summary": "Test LDAP connection",
                "parameters": [
                    {
                        "description": "LDAP test request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.LDAPTestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.LDAPSettingsRequest": {
            "type": "object",
            "properties": {
                "application_base_dn": {
                    "type": "string"
                },
                "application_filter": {
                    "type": "string"
                },
                "application_object_class": {
                    "type": "string"
                },
                "base_dn": {
                    "type": "string"
                },
                "bind_dn": {
                    "type": "string"
                },
                "bind_password": {
                    "type": "string"
                },
                "ca_cert": {
                    "type": "string"
                },
                "client_cert": {
                    "type": "string"
                },
                "client_key": {
                    "type": "string"
                },
                "default_role": {
                    "type": "string"
                },
                "display_name_attr": {
                    "type": "string"
                },
                "email_attr": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "insecure": {
                    "type": "boolean"
                },
                "server_url": {
                    "type": "string"
                },
                "start_tls": {
                    "type": "boolean"
                },
                "timeout": {
                    "$ref": "#/definitions/time.Duration"
                },
                "user_attr": {
                    "type": "string"
                },
                "user_filter": {
                    "type": "string"
                }
            }
        },
        "api.LDAPTestRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "base_dn": {
                    "type": "string"
                },
                "bind_dn": {
                    "type": "string"
                },
                "bind_password": {
                    "$ref": "#/definitions/safe.String"
                },
                "ca_cert": {
                    "type": "string"
                },
                "client_cert": {
                    "type": "string"
                },
                "client_key": {
                    "$ref": "#/definitions/safe.String"
                },
                "default_role": {
                    "type": "string"
                },
                "display_name_attr": {
                    "type": "string"
                },
                "email_attr": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "insecure": {
                    "type": "boolean"
                },
                "password": {
                    "type": "string"
                },
                "server_url": {
                    "type": "string"
                },
                "start_tls": {
                    "type": "boolean"
                },
                "timeout": {
                    "$ref": "#/definitions/time.Duration"
                },
                "user_attr": {
                    "type": "string"
                },
                "user_filter": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "model.LDAPSettings": {
            "type": "object",
            "properties": {
                "application_base_dn": {
                    "type": "string"
                },
                "application_filter": {
                    "type": "string"
                },
                "application_object_class": {
                    "type": "string"
                },
                "base_dn": {
                    "type": "string"
                },
                "bind_dn": {
                    "type": "string"
                },
                "bind_password": {
                    "$ref": "#/definitions/safe.String"
                },
                "ca_cert": {
                    "type": "string"
                },
                "client_cert": {
                    "type": "string"
                },
                "client_key": {
                    "$ref": "#/definitions/safe.String"
                },
                "default_role": {
                    "type": "string"
                },
                "display_name_attr": {
                    "type": "string"
                },
                "email_attr": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "insecure": {
                    "type": "boolean"
                },
                "server_url": {
                    "type": "string"
                },
                "start_tls": {
                    "type": "boolean"
                },
                "timeout": {
                    "$ref": "#/definitions/time.Duration"
                },
                "user_attr": {
                    "type": "string"
                },
                "user_filter": {
                    "type": "string"
                }
            }
        },
...
        "safe.String": {
            "type": "object"
        },
        "time.Duration": {
            "type": "integer",
            "enum": [
                -9223372036854775808,
                9223372036854775807,
                1,
                1000,
                1000000,
                1000000000,
                60000000000,
                3600000000000,
                -9223372036854775808,
                9223372036854775807,
                1,
                1000,
                1000000,
                1000000000,
                60000000000,
                3600000000000,
                1,
                1000,
                1000000,
                1000000000
            ],
            "x-enum-varnames": [
                "minDuration",
                "maxDuration",
                "Nanosecond",
                "Microsecond",
                "Millisecond",
                "Second",
                "Minute",
                "Hour",
                "minDuration",
                "maxDuration",
                "Nanosecond",
                "Microsecond",
                "Millisecond",
                "Second",
                "Minute",
                "Hour",
                "Nanosecond",
                "Microsecond",
                "Millisecond",
                "Second"
            ]
        },
        "util.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "err": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "util.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "err": {
                    "type": "string"
                }
            }
        }
    }
}